+++ fast/stages/03-data-platform/dev/main.tf
@@ -21,6 +21,7 @@ module "data-platform" {
   billing_account_id = var.billing_account.id
   composer_config    = var.composer_config
   data_force_destroy = var.data_force_destroy
+  data_catalog_tags  = var.data_catalog_tags
   folder_id          = var.folder_ids.data-platform
   groups             = var.groups
   network_config = {