+++ tests/examples/data_solutions/data_platform_foundations/test_plan.py
@@ -23,5 +23,5 @@ FIXTURES_DIR = os.path.join(os.path.dirname(__file__), 'fixture')
 def test_resources(e2e_plan_runner):
   "Test that plan works and the numbers of resources is as expected."
   modules, resources = e2e_plan_runner(FIXTURES_DIR)
-  assert len(modules) == 40
-  assert len(resources) == 296
+  assert len(modules) == 41
+  assert len(resources) == 313